src/utils/metrics.py
import numpy as np
from sklearn.metrics import mean_squared_error, mean_absolute_error
from typing import Dict, List, Tuple

def calculate_metrics(predictions: np.ndarray, 
                     targets: np.ndarray,
                     threshold: float = 0.5) -> Dict[str, float]:
    """Calculate various recommendation metrics."""
    metrics = {
        'mse': mean_squared_error(targets, predictions),
        'rmse': np.sqrt(mean_squared_error(targets, predictions)),
        'mae': mean_absolute_error(targets, predictions)
    }
    
    # Add additional recommendation-specific metrics
    metrics['precision'] = precision_at_k(predictions, targets, k=10)
    metrics['recall'] = recall_at_k(predictions, targets, k=10)
    metrics['ndcg'] = ndcg_at_k(predictions, targets, k=10)
    
    return metrics

def precision_at_k(predictions: np.ndarray,
                  targets: np.ndarray,
                  k: int = 10) -> float:
    """Calculate Precision@K metric."""
    # Implementation details...
    pass

def recall_at_k(predictions: np.ndarray,
                targets: np.ndarray,
                k: int = 10) -> float:
    """Calculate Recall@K metric."""
    # Implementation details...
    pass

def ndcg_at_k(predictions: np.ndarray,
              targets: np.ndarray,
              k: int = 10) -> float:
    """Calculate Normalized Discounted Cumulative Gain (NDCG) metric."""
    # Implementation details...
    pass