requirements.txt
# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0  # Ez a fontos frissítés

# Deep Learning & NLP
torch>=2.1.0
transformers>=4.34.0
sentence-transformers>=2.2.2
nltk>=3.8.1
scikit-learn>=0.24.0
anthropic>=0.3.0
pyarrow>=7.0.0


# Testing
pytest>=7.4.2

# Utilities
tqdm>=4.66.1
python-dotenv>=1.0.0
requests>=2.26.0
click>=8.0.0
pyyaml>=5.4.0
tqdm>=4.62.0